_infra/monitoring/requirements.txt
jinja2==3.1.2
python-dotenv==0.20.0
sentry-sdk==1.12.1
snowflake-connector-python[pandas]==3.9.1
premailer==3.10.0
pandas==2.2.2